+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+* Implemented `Connection::get_info()` for driver/vendor metadata ([#9]).
+
+[#9]: https://github.com/ClickHouse/adbc_clickhouse/issues/9
+
 ## [0.1.0] - 2026-07-01
 
 First stable release of the driver.